Changed Load/Save SettingAsync to be better
at handling different data types - instead of converting values to
Strings, they will be stored as IPropertyValues
Removed unused cderror header
Added skeleton class diagram
Merged CheckBoxProgressive_Loaded and CheckBoxCrop_Loaded into
ToggleSwitch_Loaded
Merged CheckBoxProgressive_Toggled and CheckBoxCrop_Toggled into
ToggleSwitch_Toggled
Changed numberProcessorsToUse to UINT32 and bytesRAMToUse to UINT64 to
better align with the saved settings types
Added more robust text to short date conversion for the privacy policy,
using std::get_time

+++ JPG Spinner/Scenario_AfterPick.xaml.h
@@ -33,8 +33,8 @@ namespace JPG_Spinner
         TypedEventHandler<ListViewBase^, ContainerContentChangingEventArgs^>^ _delegate;
 		IWICImagingFactory * pIWICImagingFactory;
 		bool CoInitializeExSucceeded;
-		unsigned long numberProcessorsToUse;
-		unsigned long long bytesRAMToUse;
+		UINT32 numberProcessorsToUse;
+		UINT64 bytesRAMToUse;
 		volatile std::atomic_ulong imagesSelected;
 		volatile std::atomic_ulong imagesAnalysed;
 		volatile std::atomic_ulong imagesToBeRotated;
